Update params.pp
when parser=future is set, operatingsystemmajrelease values should all be string. Here in this file I have updated the 7 and 5 values to string - '7' and '5'. Or else it would fail for centos 7 taking the default as the package list and try to install pam_ldap as well on centos 7 which is not required.